refactor: extract crop logic to shared utility
