+++ app/shared/components/design-system/all-components/base-card/BaseCard.tsx
@@ -9,16 +9,10 @@ import Button from '~/ds-components/button/Button';
 
 import { styles } from './BaseCard.styles';
 
+import { buildCroppedStyle, type CropRect } from '~/lib/utils/cropUtils';
 import { SvgImage } from '~/shared/components/svg-image/SvgImage';
 
 export type Variant = 'news' | 'press';
-
-export type CropRect = {
-  x: number;
-  y: number;
-  width: number;
-  height: number;
-};
 
 export interface BaseCardProps {
   image: string;
@@ -29,30 +23,6 @@ export interface BaseCardProps {
   href: string;
   variant: Variant;
   dataTestId?: string;
-}
-
-function buildCroppedStyle(
-  crop: CropRect,
-  natW: number,
-  natH: number,
-  containerW: number,
-  containerH: number
-): React.CSSProperties {
-  const scaleX = containerW / crop.width;
-  const scaleY = containerH / crop.height;
-  const scale = Math.max(scaleX, scaleY);
-  const translateX = -(crop.x * scale) + (containerW - crop.width * scale) / 2;
-  const translateY = -(crop.y * scale) + (containerH - crop.height * scale) / 2;
-  return {
-    position: 'absolute',
-    top: 0,
-    left: 0,
-    width: natW,
-    height: natH,
-    maxWidth: 'none',
-    transformOrigin: '0 0',
-    transform: `translate(${translateX}px, ${translateY}px) scale(${scale})`
-  };
 }
 
 const BUTTON_CONFIG = {
